Break DataTable component up into smaller files
