+++ ui/components/DataTable/index.ts
@@ -1,0 +1,7 @@
+import DataTable from "./DataTable";
+import { filterByStatusCallback, filterConfig } from "./helpers";
+import { Field } from "./types";
+
+export default DataTable;
+
+export { filterByStatusCallback, filterConfig, Field };
